Work on breaking out common workflow code
