+++ src/main/java/org/breedinginsight/brapps/importer/services/processors/experiment/ExperimentUtilities.java
@@ -9,12 +9,14 @@ import java.util.stream.Collectors;
 public class ExperimentUtilities {
 
     public static final CharSequence COMMA_DELIMITER = ",";
+    public static final String TIMESTAMP_PREFIX = "TS:";
 
     public static List<ExperimentObservation> importRowsToExperimentObservations(List<BrAPIImport> importRows) {
         return importRows.stream()
                 .map(trialImport -> (ExperimentObservation) trialImport)
                 .collect(Collectors.toList());
     }
+
 
 
 }
